requirements.txt
django
django-rest-framework
djangorestframework-simplejwt
psycopg2
python-decouple


